fix: Melhorar detecção de colisão com parede

+++ src/ball.h
@@ -7,8 +7,8 @@ typedef struct {
 	int x;
 	int y;
 	double radius;
-	int speedX;
-	int speedY;
+	double speedX;
+	double speedY;
 	Color color;
 } Ball;
